+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Features:
+- Add prop `closureEvents` to close the dropdowns.
+
 ## [v0.18.0] - 2016-11-14
 
 Features: